Ajuste funcionamento banco de dados

+++ Laboratorio16/app/src/main/java/br/com/impacta/kotlin/laboratorio16/adapters/ContatosAdapter.kt
@@ -46,7 +46,7 @@ class ContatosAdapter(val context:Context, val contatos : ArrayList<Contato>) : 
     }
 
     override fun getItemId(position: Int): Long {
-        return getItem(position).idContato
+        return getItem(position).idContato.toLong()
     }
 
     override fun getCount(): Int {
