+++ Laboratorio16/app/src/main/java/br/com/impacta/kotlin/laboratorio16/models/Contato.kt
@@ -1,7 +1,7 @@
 package br.com.impacta.kotlin.laboratorio16.models
 
 data class Contato(
-    val idContato : Long = -1,
+    val idContato : Int = -1,
     val nome : String,
     val telefone : String,
     val idade : Int) {
